Moved PCX code to Library/Image
