+++ src/Engine/Graphics/CMakeLists.txt
@@ -25,7 +25,6 @@ set(ENGINE_GRAPHICS_SOURCES
         OpenGL/TextureOpenGL.cpp
         Outdoor.cpp
         Overlays.cpp
-        PCX.cpp
         PaletteManager.cpp
         ParticleEngine.cpp
         PortalFunctions.cpp
@@ -64,7 +63,6 @@ set(ENGINE_GRAPHICS_HEADERS
         OpenGL/TextureOpenGL.h
         Outdoor.h
         Overlays.h
-        PCX.h
         PaletteManager.h
         ParticleEngine.h
         Polygon.h
@@ -80,7 +78,7 @@ set(ENGINE_GRAPHICS_HEADERS
 add_library(engine_graphics STATIC ${ENGINE_GRAPHICS_SOURCES} ${ENGINE_GRAPHICS_HEADERS})
 target_check_style(engine_graphics)
 
-target_link_libraries(engine_graphics libluajit glad nuklear utility library_serialization library_color)
+target_link_libraries(engine_graphics libluajit glad nuklear utility library_serialization library_color library_image)
 if (NOT BUILD_PLATFORM STREQUAL "android")
     find_package(OpenGL REQUIRED)
     target_link_libraries(engine_graphics ${OPENGL_opengl_LIBRARY})
