+++ src/Library/CMakeLists.txt
@@ -6,6 +6,7 @@ add_subdirectory(Color)
 add_subdirectory(Compression)
 add_subdirectory(Config)
 add_subdirectory(Json)
+add_subdirectory(Image)
 add_subdirectory(Lod)
 add_subdirectory(Logger)
 add_subdirectory(Random)
