+++ src/Library/Image/PCX.cpp
